requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
streamlit==1.32.2
langchain==0.1.13
langchain-community==0.0.29
langchain-groq==0.1.0
langchain-huggingface==0.0.3
sentence-transformers==2.6.1
faiss-cpu==1.8.0
pypdf==4.1.0
python-multipart==0.0.9
requests==2.31.0
python-dotenv==1.0.1
